requirements.txt
torch
torchvision
pillow 
numpy
transformers
requests
streamlit
style 